Improved: Improve ObjectInputStream class (OFBIZ-10837)
Cleans and simplifies things in UtilObject.java and also handles patterns. 
That's what we missed most when needing to update.

Fixes typos in SafeObjectInputStream class and add a commented line for a WIP in
resolveProxyClass method

Also includes work done in UtilObject.java for OFBIZ-9855 
"Using try-with-resources with File IO Objects."




git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/trunk@1857392 13f79535-47bb-0310-9956-ffa450edef68