Improved: Update build.gradle to the latest dependencies (OFBIZ-10213)
At https://github.com/ben-manes/gradle-versions-plugin Ben Manes suggests to 
have a look at few Gradle plugins which depends on his own plugin. 

I checked and there is one which is interesting: 
https://github.com/patrikerdes/gradle-use-latest-versions-plugin. 
It does much of the work you would have to do manually when updating libs. 

Beware though that this is only a help. If you use it without check you will 
need to check things by yourself (can be as tedious as not using this plugin)


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/trunk@1857599 13f79535-47bb-0310-9956-ffa450edef68