Fixed: Wrong permission check in ProductScreens#FindProduct (OFBIZ-10925) Thanks Dennis Balkir for your contribution
The permission service acctgAgreementPermissionCheck is used in the screen FindProduct in ProductScreens.xml. 
This doesn't really make sense, since there is no connection to the accounting application here.
There was the service productGenericPermission which could have been used instead, 
but this one calls the simple-method checkProductRelatedPermission, 
which then check for the CATALOG permission. Since this is already done in the CommonProductDecorator before, it would not make any sense to do this a second time.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/trunk@1857617 13f79535-47bb-0310-9956-ffa450edef68