Fixed: If the picklist cancelation process failed, the picklist item shouldn't be marked as cancel (OFBIZ-4690)
When a transaction is running in which Picklist entity is updated (i.e. canceled), now if the transaction is rollback the picklist status will not be changed but as per previous code, the picklist item's status will get updated i.e. marked as cancel because we called the cancelPicklistAndItems service in async mode

So the situation here was the Picklist is not canceled, but the Picklist items were marked as cancel.

Thus to fix it, call the cancelPicklistAndItems service in sync mode.

Thanks:  Jacques Le Roux for your thoughts.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/trunk@1857813 13f79535-47bb-0310-9956-ffa450edef68