Fixed: Update invoice item looses invoice context (OFBIZ-)
When clicking  the "update" button on an invoice item the context to the 
invoice (invoiceId) is lost. The result is an empty form to add a new invoice 
item without invoice context.

Go to /accounting/control/listInvoiceItems?invoiceId=demo10001
Click on "update" below the existing items. 
You will then be redirect to /accounting/control/listInvoiceItems 
(without invoiceId). You will not see the invoice items as the context to the 
invoice is gone.

jleroux: the same existed for budget items. As Ingo mentioned:
<<It was broken by OFBIZ-9997:
Replace request-redirect w/ no redirect-param attribute 
by request-redirect-noparam>>

Now we need to check all changes done for OFBIZ-9997

Thanks: Ingo Wolfmayr for report and fix proposition

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/trunk@1859033 13f79535-47bb-0310-9956-ffa450edef68