Fixed: field emplFromDate is forgot in PayHistory entity (OFBIZ-11028)
Arpit Mor: steps to regenerate:
    Login to the URL: https://demo-trunk.ofbiz.apache.org/humanres/control/main
    Click on Employments
    Click on New Employments
    Click on Create
Actual: Error message is displayed.

Olivier Heintz: employment is associated with PayHistory, and there is a problem 
in PayHistory entity definition in OFBiz,
In PayHistory, the field fromDate from Employment is confused with fromDate 
about the current record.
A Employment can have multiple PayHistories and should have multiple
because PayHistory should show history of Pay  for a employment  !
It's necessary to have a field emplFromDate (to have the complete employment 
primaryKey).
When modifying a PayRecord the current should be expire and a new one should 
be created.

Thanks: Arpit Mor for report at OFBIZ-10969 and Olivier Heintz for the fix

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/trunk@1859571 13f79535-47bb-0310-9956-ffa450edef68