Fixed: Drop-ship process behaves incorrectly when a combination of drop-ship and non-drop-ship products are added into the cart. (OFBIZ-11021) RequirementMethodEnumId should be decided in the order of property set at ProductStore -> Facility -> Product for each order item which was not working if the non-drop shippable product(s) is added after the drop shippable product(s).
Original Flow: requirementMethodEnumId was initialized outside the loop and once it is initialized with Product level information it never gets re-initialized if Product is non-drop shippable.
We have updated the code to use the storeRequirementMethodEnumId variable for store level information which is defined outside the loop and this will be used to initialize requirementMethodEnumId for each order item and then product level information can override it. That means for each product ProductStore -> Facility -> Product this rule will work properly.

Thanks Pawan Verma for reporting the issue and providing the patch and Jacques Le Roux for review.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/trunk@1859909 13f79535-47bb-0310-9956-ffa450edef68