Implemented: Remove the Gradle wrapper jar (OFBIZ-10145) (LEGAL-288)
As regards to LEGAL-288 [1] we remove the gradle-wrapper.jar from the source and future release packages.
Now when you run ./gradlew (or gradlew), we check before that gradle/wrapper/gradle-wrapper.jar is present
otherwise we download it before calling the wrapper.

To realize the download, we need as a prerequisite to have:
Unix-like: curl or wget
MS windows: wget

If any issue appears, it's possible to set gradle-wrapper.jar and gradle-wrapper.properties by hand on the directory gradle/wrapper/
Thanks to Jacques Leroux and Swapnil M Mane for their help

[1] https://issues.apache.org/jira/browse/LEGAL-288

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/trunk@1861766 13f79535-47bb-0310-9956-ffa450edef68