Fixed: Issue of unable to view a PartyContent on view profile page of a party (OFBIZ-11038)
When click on view button img request will be hit along with contentName and imgId as a parameter. This request invokes serveImage event(DataEvents.java) which is deprecated now.
Instead of img request we can call stream request with contentId as a parameter that invokes serveObjectData event.
This will work as per the value of content-disposition-type (requestHandler.properties) i.e. attachement or inline

Thanks: Devanshu Vyas for reporting the issue and Humera Khan for the patch and Prachi Shastri for testing of the bug.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/trunk@1865852 13f79535-47bb-0310-9956-ffa450edef68