Improved: Give nested requests related with checkLogin a propper Error-Message in log (OFBIZ-11154)
An error appears in log due to the nested request that can not be resolved 
properly.  

We should not worry about nested requests, they continue to work, and we should 
simply not show an error message from RequestHandler in case of nested request.

Thanks: Initialy spotted by Benjamin Jugl in OFBIZ-9379

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/trunk@1865902 13f79535-47bb-0310-9956-ffa450edef68